refactor: AppBaseModel 정의를 공통 모듈로 분리해 임포트
